server/requirements.txt
fastapi>=0.100.1
uvicorn>=0.23.2
python-multipart>=0.0.6
python-dotenv>=1.0.0
whisperx==3.3.3
torch>=2.0.1
numpy>=1.24.3
pydantic>=2.1.1
ffmpeg-python>=0.2.0
tqdm>=4.66.1
google-cloud-storage==3.1.0